+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.4] - Unreleased
+
+### Added
+- Run docker cronjobs in EC2
+
+## [0.0.3] - 11/10/2020
+
+### Added
+- Deploy script
+
 ## [0.0.2] - 11/10/2020
 
 ### Added